requirements.txt
streamlit==1.24.1
Pillow==9.3.0
transformers==4.34.0
torch==2.0.1
